move components into containers
