+++ web-app/stories/Continue.stories.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
 
-import Continue from '../src/components/Continue'
+import Continue from '../src/containers/Continue'
 import demo from './data/basic'
 
 storiesOf('Continue', module).add('Page', () => <Continue tutorials={[demo]} onContinue={action('onContinue')} />)
