+++ web-app/stories/New.stories.tsx
@@ -3,6 +3,6 @@ import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { action } from '@storybook/addon-actions'
 
-import New from '../src/components/New'
+import New from '../src/containers/New'
 
 storiesOf('New', module).add('Page', () => <New onNew={action('onNew')} />)
